Compute per method statistics
This CL computes and saves the inclusive and exclusive times
per method.

The inclusive time of a method is a time spent in that method
and all of its callees. The exclusive time is the time spent
in only in the method itself, ignoring the time spent in its
callees.

We also keep track of methods that are recursive and compute
their inclusive times appropriately (i.e make sure we don't
double count the inclusive time for recursive calls).

Change-Id: Ibaaca78c2454ca39ab1e6e581cb7019ff8cabd6f