Fix eclipse import handling of Java libraries
This changeset fixes handling of dependencies where you have
a library project depending on non-Android libraries (this is
not recorded in project.properties; it is only handled in the
.classpath file as a workspace relative path).

This changeset also cleans up path handling from the .classpath
file and improves resolution of workspace paths.

I will add tests to this shortly; I've been testing these changes
manually with a complex project; I just want to get this CL in
in time for 0.4.

Change-Id: I45cb22a6429df4c92ad5f8c47c45b0abc1f65af3