Add PackagingOptions on libs and fixed test issue with local jars.
packagingOptions on library projects allow controlling what goes in the
test app only (the library itself does not pack anything).

To solve the issue with local jars in libraries, we use the same
mechanism that Ant/ADT uses and filters duplicate jar files by
sha1. This is temporary while we wait to be able to do a proper
dependency between the test app and the tested lib.

Change-Id: Ibb48c803fe0f0f317d0dd0a21d145d5fdf357749