Add lintOptions#checkReleaseBuilds which runs lint on release builds
If lintOptions.checkReleaseBuilds is true (which it is by default) the
assemble task for non-debug targets will now run lint and check all
issues that have severity fatal.

It also tweaks the error message shown when the build aborts due to
a lint error. If you run a lint target, the message will look like this:

> Lint found errors in the project; aborting build.

  Fix the issues identified by lint, or add the following to your build
  script to proceed with errors:
  ...
  android {
      lintOptions {
          abortOnError false
      }
  }
  ...

If the build fails during release target assembly, it looks like this:

* What went wrong:
Execution failed for task ':lintVitalRelease'.
> Lint found fatal errors while assembling a release target.

  To proceed, either fix the issues identified by lint, or modify your
  build script as follows:
  ...
  android {
      lintOptions {
          checkReleaseBuilds false
          // Or, if you prefer, you can continue to check for errors
          // in release builds, but continue the build even when
          // errors are found:
          abortOnError false
      }
  }
  ...

Change-Id: I87debf8df0ed88f327b4df7119a9dfb057c34945