Fix lint format detector for xliff strings
For xliff formatting strings we replaced the placeholder elements with
the parameters instead, e.g. for the string

<string name="name">
   First Name: <xliff:g id="firstName">%1$s</xliff:g>
instead of showing
   First Name: %1$s
we show
   First Name: ${firstName}

However, this was interfering with for example lint's formatting
string argument checker, which no longer found the formatting strings
(%1$s) to count them.

This changeset fixes that bug, and introduces a new method on
layoublib's ResourceValue class: getRawXmlValue() which returns the
raw XML string in a string declaration (which will let you for example
also find back the markup string content, such as <b>bold</b> tags
etc.)

Change-Id: I6a7ad307fbc86bface00afa8a18d970226dc3db0