Update ViewInfo to return the type of view.
The type of view was earlier obtained by using a differnt type of
cookie. This change allows obtaining the view type in addition to the
view cookie, which can still point to the XML tag.

The method removed from ViewInfo hasn't been added to any released
LayoutLib yet. Hence, it is okay to change it.

Change-Id: Ib79bccaac92d371f4ef6ac201a556a86ff85e2bf