fix for 69782 : compatible-screen handling incorrect.
We used to consider the screen elements of the compatible-screens section
to be keyed by the screenSize attribute. This is incorrect, each screen
element is identified by a combination of the screen size and density.

With this change, all compatible-screens elements will be merged
automatically into the higher priority document (unless tools:remove
or tools:removeAll are used).

Change-Id: Iee5f9af819622cf42cf5625e1e0c9f850bbe37a0