Warn about plural strings that do not contain formatting arguments
In for example English, the plurals "one" quantity maps to a single
value: "1". Therefore, it's okay to write a message for this quantity
which does not include the actual number, since it's implied.
However, in many locales, quantities which may seem to apply to a
single number (zero, one and two) may actually match multiple
values. For example, in Slovene, the "one" category will match 1, 101,
201, 301, and so on. In these cases it is normally an error for the
message to *not* actually include the quantity, so that's what the new
lint issue checks.

This CL also extracts out the plurals database from the detector into
a separate class (which a separate CL builds on to replace the
hardcoded list with a parser which uses the ICU rules file directly).

Change-Id: Ib2c2d0fb220b90cff546889f527f06b5f8355313