SDK loader: mark build-tools as broken if they don't match the current OS.
This loads the build-tools as "broken packages" so they are
visible in the SDK Manager but they are not functional as
actual build-tools.

This is only useful for users who manually create build-tools
local packages -- when installed via the SDK Manager, the OS
is checked before download so it's not technically possible for
a user to download and install an invalid build-tools package.

Change-Id: I18e71de271cf02a93708d0204c9d26f0f05d5bce