Migrate from packageName to applicationId in Gradle DSL
This CL replaces packageName, packageSuffix, testPackage and
testedPackage with applicationId, applicationIdSuffix, etc.

This helps decouple the concept of the manifest package, which
controls where the R class is located and how relative manifest
classes are resolved, from the concept of the immutable application
id which should never change once an application has been published
(since otherwise you cannot upgrade from one version to another).

Once this is in, Studio refactoring can stop omitting the package
attributes in manifests (as long as the applicationId is defined).

The various package-named DSL methods continue to work, but emit
build warnings.

Change-Id: I0c3a32394af3f63c82f9d4c0b2562dc601195429
69679: Rename packageName to applicationId in defaultConfig