Update mobile template to be aware of Wear
If the user selects both wear and mobile from the FormFactor page, then
the mobile app will automatically bundle the wear APK.

Known issue - The following action path will result in a sync error.

Select Phone/Tablet and Wear from the form factor page
Continue to the end of the wizard, but don't hit Finish
Go back all the way to the form factor page
Unselect Wear
Select Glass or TV
Continue to finish the wizard

In this case, since the wear module was selected and all the variables
for it were set (by going through the add activity flow), the mobile
app will still believe that there is a wear module that will appear in the
project and gradle sync will fail because the module dependency cannot be
resolved.

Fixes b/15777961 and b/15806401

Change-Id: Ic4a3b3d86512f0a065295704401f7803ac4e63fb