Density based multi-apk.
The new model replaces using ProductFlavors for
doing density-specific splits. This new mechanism is more narrow
focused and there is in fact a single variants with multiple outputs
(one for each density + a universal one).

This means all the outputs have the same source folders (and share a single
compilation task, same for dex, proguard, etc...), the same dependencies, etc...

To enable this new splits, the following DSL can be used:

  splits {
    density {
      enable true
      exclude "ldpi", "tvdpi", "xxxhdpi"
    }
  }

Also available:

  splits {
    density {
      enable true
      reset true // remove the existing density list.
      include "mdpi", "hdpi", "xhdpi", "xxhdpi"
    }
  }

Change-Id: I844f9f35196133a7d42947cc721b13ceb4ed27d9