Add API to get multiple matching resources.
The API enables ignoring specific qualifiers in a FolderConfiguration by
setting them to null. This means that multiple resources match a given
FolderConfiguration, all of which may differ only in the ignored
qualifiers.

This is important for ignoring the API levels in the layouts, thereby
enabling layouts meant for old API levels to be rendered by newer
rendering library.

Bug: http://b.android.com/75156
Change-Id: I4ad8559a111ea6cfd49b8783d4af503fd2bd8a34