Fix race condition in lint's resource folder cache
Lint was caching the folder version for a given parent folder.
However, this is problematic when lint is run from Gradle (where tasks
can execute in parallel), and in particular one large project was
seeing intermittent build errors as a result of this, where it was
evident that the build was seeing an inconsistent view of the parent
folder cache.

Therefore, simply move the computation and caching into the
LintDriver, which is always unique for each lint job, rather than
storing this data in static fields.

Change-Id: Ic1ab9d08af1b6e73d1f38b9c6fd6487cc3b38fba