generate error when AndroidManifest.xml is not parsable.
early on during the project configuration, we use xpath to extract
the package name, version and few other attributes from the
AndroidManifest.xml. If this file is not XML parsable for any reason,
no real failure was generated leading to obscure failures later on.

Generate a build failure when the xpath code cannot get the package
name from the AndroidManifest.xml which is mandatory.

Change-Id: I8e0f8ad0016750bceb7d0abcf6a447ba03654831