Enhance error message when dependent project are not configured correctly
with a project with multiple modules and one module is not present or empty,
the build will fail with a message :
Configuration with name 'default' not found.

which is not quite clear which one is the culprit. added some context
information to the message :
Cannot evaluate project client-common-java : Configuration with name 'default' not found.

Change-Id: Iec6f3a4d7a2d4bceb5f4734ff2b6e6784ebd4b4d
(cherry picked from commit 68808e88166b0ac11765750716018476d8e3763c)