force density ordering
Density used to be stored in a vanilla Set to not have duplicates but these
densities are used to be build the list of VariantOutput where we expect the
universal one to be the first in the list.
This created issues after switching to JDK8.

Change-Id: Ia460c7736fd6d489ca95bfdac02f3dffa8dbacd5
(cherry picked from commit 1e08713f8b45684f490f23853337aded6dcb48c2)