Adds lint check for signatures related security exploit.
The issue here is that when examining signatures from
PackageInfo, if a developer isn't careful, they could
forget to validate all the signatures returned and
instead only check for the presence of a single
certificate they're expecting--opening themselves
up to the multiple certificate exploit.

More info:

https://bluebox.com/technical/android-fake-id-vulnerability/

We add a lint check to warn for this case.

Change-Id: I557d5c9ede4e6afef7578c069dabf25125b8b15a
Signed-off-by: Shannon Chen <shannonchen@google.com>