Add support for BCP-47
As of Lollipop, Android supports BCP-47 as an alternative to
specifying the language and region using 2 letter language codes.  The
syntax is using a language tag as specified in
   https://tools.ietf.org/html/bcp47
but with the prefix "b+", and with "+" instead of "-" to separate the
segments in the BCP-47 language tag.

In this CL the various code used by the Gradle plugin, lint and the
IDE are updated to handle these new tags. The biggest change for
clients is that they should no longer rely on looking up the langauge
by simply consulting FolderConfiguration.getLanguageQualifier(); the
language can now be alternatively specified with a LocaleQualifier, so
instead, use FolderConfiguration.getEffectiveLanguage() and
getEffectiveRegion() to look up the appropriate language or region, as
necessary.

This CL also adds in various unit tests, for example making sure that
the lint translation checks catches a missing translation in Klingon,
which is now a valid language tag...

Change-Id: Ia92f09870b3f113c94cc855598007f567e1dcf1c
(cherry picked from commit 25caa30fbcb1ff4bb78f84658c65f0edb82c659b)