draw9: Move help panel out of the image viewer.
The N patch viewer contains a image viewer which is the main component
where the N patch is edited. This component has a JPanel inside of it
along with its custom rendering canvas. This causes issues with keyboard
focus control (focus always goes to the panel and never to the canvas).
In order to get the default focus manager to work (rather than implementing
a custom focus traversal policy), this CL just moves the jpanel out of the
image viewer.

Change-Id: Ib45c0379d5910296f9bb377757185e14933c0918