140267: Fix NPE in API for test variant.
The setter for the tested variant was called after the test variant was
added to the collection. This means that any closure setup by the developer
in the build.gradle file would be called before the setter and therefore
the testVariant API would return null on getTestedVariant.

Change-Id: I83e7ec27966db3fef5d761440f3bd621fefd3c89
(cherry picked from commit 7e553cb00d86bd023deb50639294da6956a627c3)