reverted a static compilation change from TaskManager
When we introduced the static compilation, a potential bug in the gradle code
surfaced (still chasing the reason on the gradle side) and the archivesBaseName
settings was not accessible any longer.

Reverted to use dynamic invocation for this property only until a fix is
eventually provided by Gradle.

added a regression test

https://code.google.com/p/android/issues/detail?id=155888

Change-Id: Ib47f7d59c89e7647e846392289147338a56d98fe