Properly handle deps with classifier in the androidTests.
If a project depend on an artifact in 'compile' and
the same artifact (but with a different classifier) in
'androidTest', the code removing duplicate would fail
to take into account the classifier, removing the dependency
completely.

This fixes this by using groupId:artifactId:classifier
as the key when searching for similar artifacts.

Change-Id: I3505dbf47b7c8d4b50b21bd50c842ee7ce8b20df
(cherry picked from commit e695621cd0d23e0a4e6a9a8724f17a3df64173fe)