Refactoring pointer size in perflib.
Moved pointer size and all type sizes to Snapshot, as pointer size is
dependent on the Snapshot used. The old bug was that if the pointer
size remained static, then multiple snapshots opened at the same time
(when we move to 64-bit pointers) will cause parsing and/or data
corruption.

Change-Id: I0696dc59228b25268e8ec5ffbfe4735903bbb8c3