ddmlib DeviceMonitor: Fix synchronization around channel registration
Whenever a new device is detected, a socket channel is opened to
adb to listen to the VM's on that device. This channel needs to be
registered with the selector that multiplexes across many such
channels.

This CL cleans up the synchronization between the two: the registration
of the new channel now happens on the same thread where the selection
happens.

Change-Id: I214b245aade156cae20b845a0eadf4d07ecae417