Removes ability to change code before constructor super call
Code before super can use "this" in limited cases and it cannot be used
to send to another object. When a non static inner class is used, variables
of the outer class can be referenced through the uninitialized "this" before the
super call. These references cannot be easily accessed from the outside
without a reference to the uninitialized this.

Change-Id: I52d5dda2de412b1a70bb1c56caac18b863a5e7e2