Transform API.
During setup of the tasks, we often have cases
where optional tasks consume and transform elements.

This makes it difficult for the downstream tasks to
be properly configured as they have to handle all
possible combinationis upstream. For instance having
or not having jacoco, proguard, multi-dex, etc...
complicate things.

The goal of the transform API is to automate this
for a subset of the types handled by the build
(classes, dex, java resources).

A Transform indicate which type and which scope it
applies to and how it transforms them. The
TransformManager handles adding transforms, creating
intermediary content objects between transforms and
setting up dependencies based on what generates the
input content objects.

This is a first commit creating the basic API, and the
manager and custom Task type handling Transforms.

Change-Id: I8c1adcc89a1ce595d65cf1c85b0221a2a8062d91