Create the new NavigationViewActivity using material design widgets.
The layout and Java code for using material design is separate from
the older version. The new approach is in separate folders:
   res-v22 and src-v22.

We are reusing the App Bar and the simple Relative layout from common.
If no App Bar is wanted use a Toolbar which has better support for
an ActionBarDrawerToggle.

Change-Id: If8b5d6c574ae4e79bfaabf929f7fe195ffde30fb