Change the MasterDetailActivity wizard to generate a toolbar.
Add a toolbar such that we can place a FloatingActionButton on
the master, detail, and combined views.
I have chosen to NOT have the AppBar resize when looking at the
master and the combined view.
The list fragment does not support nested scrolling and neither
does LinearLayout (for the combined view).

Change-Id: Icb93930872fb8926f023996f2a83fee401c8cbb6