Fix test library dependencies.
Local jars of a library are showing up as direct
dependencies of the test artifact as well (They are
also showing as the local jars of the aar that represent
the library itself), causing duplicated class errors.

In order to fix this, we override the behavior of the
custom AndroidLibrary object that represents the tested
library to return an empty list of local jars.

Change-Id: I26647dc52848fdd16cd0958920278539e50a4fa7
(cherry picked from commit a9cabb0d568b616b747bcefc82269ef266ae54c9)