Put back getDex() in ApkVariant.
This throws an exception with a message explaining the Transform.
This is better than having an error on a missing method.

We should put a link to the javadoc of the transform, and
how to register them.

Change-Id: Ifa70087a48dd08e7265c9c2dfaf27cb9a8ad5c26
(cherry picked from commit 54d01e927f9ec12ec66497f59ff4ee779abc57a5)