Don't flag hardcoded string warnings on place holders
Our templates, and our layout editor, creates a few
place holder strings (e.g. "New Button", and
"Hello World!".) Don't flag warnings on these since
developers are unlikely to leave them in real apps,
and it's not friendly to create a new project that
lights up warnings in the first file we automatically
open for them.

Change-Id: I7409029b17802174d06b4f4048229bc896530878