Properly filter out R/Manifest class from AARs
If a transform generates a Jar then the standard flow
of creating a jar from the folder with a set of excludes
won't be applied.

So this sets up a copy of the generated jar with a set
of filters to skip some classes.

Change-Id: Ic34b8efd7c70992cad4acd1a2e8db33c24c7f936
(cherry picked from commit fffa955a60f1a1a21be91d35666939a2ac258de4)