fixed private field access.
When field is private, we should be looking for its definition starting at the defining class
(owner) and up rather than just use the receiver (this) class.
This allow for class hierarchies which override a field name (potentially with a different type).

Also disabled bridge method redirection since it may confuses the access$super and access$dispatch
implementation and it contains no user code.

Change-Id: Ib6da00e24a6f3a27db2f8b477e247f8d1a96fc7b