allow setting public final fields in constructor.
when setting a public field, we don't use reflection to keep things
as fast as possible. This is however an issue when the PUFIELD
used to be located in the <init> method and is now located in the
init$body. If the public field is final, we will get an illegal
access violation since we try to set a final field from a method
which is not a constructor any longer.

force use of reflection is that case.

Change-Id: I53358783c660cc53f4bdb21292cc5d8344e922b6