Fixed jar in assets bug.
Changed unzipAar file to check the jar location before copying them.
Jars in asset or res directories are not touched. classes.jar and
lint.jar are moved to the jars directory. Jars in libs are moved to
the jars directory. All other jars are ignored and a warning is issued.

Added integration test to verify jar copying.

Bug: http://b.android.com/183480
Change-Id: I399ddc466dada45ac1390f2b1fdc2421291bfd76