Do not produce the restart.dex until asked or verifier failed.
Up to now, restart.dex and reload.dex were always produced.
Now the restart.dex will only be produced when the verifier indicated
that a restart is necessary. In such a case, the reload.dex will not
be produced.

the restart.dex will contain all changes that accumulated since the
last restart.dex was produced (or the main APK).

Change-Id: Ice97813cd6476dc8bd661f92c4842dcbe417c3c9