Revert "Add help for template "package name" field"
The help message wasn't harmful per se, but actually it turns out it doesn't normally show up because of an issue with the Swing component that represents it, so might as well not clutter up the template for now. Also, there are lots of templates that are missing this help tooltip, so if we fix it here, we should fix it everywhere.

This reverts commit 8dd5c0322d8164476490d42e7d6617f02749daee.

Change-Id: Ib7b03a6b0a494a8db026190cff1e13e1004cf62c