Fix race in Namespace.
Namespace is a global map used to register BinaryClasses. It is
also used to register instances of Dynamic classes, so modifications
can be made by multiple threads at once. Changed to ConcurrentHashMap<>.

To consider, I think that correctness this should be linked to the
connection, rather than a global.

Maybe b/25709475

Change-Id: I2dd7b5e39b52c31400fb78673a29f6a9f6d51039