Fix CleanupDetector to correctly handle reassigned variables
The CleanupDetector tracks local variable assignments
such that it properly handles dataflow where you allocate
the resource and assign it to one variable, then the value
flows via variable assignments to some other variable, and
then the resource is disposed on that other variable.

However, this was not accounting for a scenario where a
value flows to a different variable, but that other variable
is reassigned some other random value.

This also updates the commit transaction unit tests such that
they don't contain any parsing errors, and turns off the
compatibility allowCompilationErrors mode for this test.

Change-Id: Ifd8b9c70500e7c96d2877b11999685e32c210e9a