Removing single static instance from AndroidSdkHandler
Before there was (in almost every case) a single AndroidSdkHandler
instance, saved statically within AndroidSdkHandler. Unfortunately
in IJ with the android plugin we need to support SDKs per module,
so this is impracticable.
Refactoring such that all config information is statically shared
between AndroidSdkHandler instances except the local path. A cache
of AndroidSdkHandler instances is maintained per path.
At this point we just key off a single empty file. In a following
change getInstance() will be made to take a file argument, and this
will be adopted across the current usages (primarially by keeping
and AndroidSdkHandler in AndroidSdkData instead of (though initially
alongside) the current LocalSdk instance).

Change-Id: Ic3577abc817c4f7c6c77d2c24a4719e4a356e616