Fix NodeUtils#updateNamespace when there are multiple NS
Previously, the code would check doc.getAttributes() for
existing xmlns:prefix attributes, and when missing, would store
newly made xmlns:prefix attributes into the first child's
attributes -- different maps.

doc.getAttributes() would often be null, which will cause the
code to use "ns1" as the new prefix. These two issues
combined would cause the code to use ns1 for any prefix,
which would be a problem if there is more than one.

Also, check for the "xmlns:" prefix in the attribute name
instead of in the namespaceURI.

Add a duplicateAndAdoptNode method to test this. adoptNode
modifies the original node, so does not allow checking
the original nodes. duplicateNode doesn't update namespace
and prefixes.

Change-Id: I92f6691a38b79c930f8401db6d7cc956690611cd