added a more elaborated secondary input notion to Transforms.
the current secondary input all provoke a non incremental execution
of the transform when changed/added/removed.

with the new mechanism, Transform can declare secondary inputs that
can be handled by an incremental execution. Changes associated
with such secondary inputs are now passed to the transform() call
to notify the Transform of incremental changes.

Change-Id: Ic8fd1f337bf912d9b3876be8667a39d870cfe4aa