write after each artifact production to a tmp build-info
when build fails in the middle (due to a compilation error for instance)
the build-info.xml was not written. Yet some artifacts like resources
may have already been produced and we then forgot about including
those once the build is successful. We now use a tmp-build-info.xml
to write temporary artifacts and we reload from it if the file is present.
when the build is successful, we delete the tmp-build-info.xml

Change-Id: Ic3b6224140cfe50e0ee8d8952cc447d6fe99ab4c