Making gradle refresh sdk each sync
Previously, a new SdkManager was created each time a gradle sync was
done, and this caused the local sdk to be re-parsed. When we changed
over to AndroidSdkHandler/RepoManager, caching was added, but gradle
relies on the sdk being re-parsed each time (since there's no
mechanism for studio to notify gradle that an update to the SDK has
occurred.
Fixing it so the sdk isn't cached across syncs.

Change-Id: Ie21985355a5979dbc73b966fcf73a53bf91fb897
(cherry picked from commit 510fe264db68dd5e096be9a94529092c411f0ed9)