switched reload.dex transform from just incremental tasks to also full
The reload.dex task used to only be ran when invoking the instant
run enabled incremental task. This was problematic when a full build
was invoked preventing the task from cleaning its state. The dirty
state could confuse the next iterations into not running the task
or prodiding outdated input information.

Change-Id: I03777ca3c08d6279b57636e3ed0732cf50fb842b