200405: First value resource change does not work
This CL fixes a bug where value resources referenced from a layout
were not always updated on warm swap.

The root issue was that as of Lollipop, there's an object pool in
the Resource manager which caches TypedArrays, and these were pointing
to pre-warm swap resources.

This CL drains the object pool as part of the resource monkey patching.

Change-Id: I01e0e56d6923d888943a52b9c21060faf3dacdb3