always force re-dexing when changes occured.
in case classes were removed and re-added, because we consider
class removal a hot-swapable change, and no other change occured,
Gradle would not rebuild the slice dex file as its input appeared
to not have changed. Adding a buildId to the slice output directory
makes sure the corresponding slice.dex is rebuilt and registered
as a produced artifact.

also do not keep removed classes events if the class was re-added
in a subsequent iteration.

Change-Id: I1f6a448c0763edc3620da398aed9fb3ab202255f