Original streams should all be lazy suppliers.
It's important to not reject missing files during evaluation
because they may not be built/downloaded yet. We should only
reject just before execution.

So, since Gradle still doesn't run convention mapping only
just before the task is run (but also at evaluation time),
it's important to have this be dynamic even if it means it'll
be run twice.

Bug: 197844

Change-Id: If1e8e386eb499ad35018963cef23e174cd0cd32a