Fix null API version in AVD manager.
A "null" can be shown for the API level in the
list of AVDs if the AVDs are edited in a certain way.
See the bug for details.
The problem was that during the "load" of the AVDs
we determmine that the device hash has changed,
which will trigger a save with some unwanted
properties included.

Bug: http://b.android.com/200908

Change-Id: I466104a66ee9691d76c663fd9c837f40f889bf81
(cherry picked from commit 0a924114b126ccc0410f9aca32f3468b31799ef4)