Fixing xml schema caching
Previously we cached schemas based only on the actual xsds that went
into them. However, the LSResourceResolver is also kept in the schema,
and must be part of the cache key (in practice it's not important, since
the resource resolver will always be initialized in the same way, but in
unit tests it can sometimes be different).

Fixes https://code.google.com/p/android/issues/detail?id=201166

Change-Id: Ibef028bfdf0b7d37ec70541eb25d0ac06124f02e