APK Signature Scheme v2 support
This makes the build pipeline sign APKs using APK Signature Scheme v2,
in addition to the usual JAR Signature Scheme (aka v1 scheme).

The only change to v1 signing is that the .SF file's header section
now contains an additional header which indicates that the APK was
supposed to be signed with v2 scheme. This is to detect v2 signature
stripping attacks.

A follow-up CL will make it possible to disable v2 signing, probably
via a new attribute in SigningConfig.

Useful testing can be added only once there is code to verify APKs
signed with APK Signature Scheme v2. A follow-up CL will add such code
because we'll need it for the apksigner tool. That CL will also add
thorough tests for signing and verification of v2 signatures. Some of
these tests already exist in CTS and thus can be easily ported over.

Bug: 26516150
Change-Id: Iccf06e485900e9c5d6cb4322596f131678844c44