200872: With 'provided' dependencies we get MissingRegistered lint check error.
If you're running lint on *just* a library, don't flag missing
classes. There are cases where this is desirable (e.g. see the
request in issue b.android.com/191987), but unfortunately, when
you run "./gradle lint" on a whole project, gradle will not just
run lint on the main app with all its dependencies, it will run
lint on each library too, in isolation, and this will then break
the build if it sees a library that references a class that has
not yet been provided.

Also clarify the issue explanation text as requested in issue 201338.

Change-Id: I7bfc83cb4fe2923d6f5d162fc5ab9c5a15add0de