Refactors animation to be separate from the paint cycle.
Animated components used to do the animation step right
in the paint cycle, which is incorrect.

The new approach is based on a centralized Choreographer
that animates nodes (arbitrary nodes) in the EDT separate
to the paint cycle. This allows for proper sharing of
interpolated values across different components.

Adds a first basic path of LineChart component, a rewrite
of TimelineComponent, that only takes care of the actual
central part of the graph, as the remaining elements are
being split up into their own components.

Change-Id: I984a661a8413795f2aa4c669eff9b11440867a04