Change installers to install listeners where possible
Previously, when we wanted some special actions to happen during sdk
component installation, we created a new subclass of PackageInstaller.
However, it turns out that we need more complex composition of behaviors
than this allowed: for example, we want to be able to apply diff-based
updates to maven packages, etc.
Change-Id: I777d6043a7f910ab86246fbb6f34b2fad1ce598d