Show a toast in the app if hotswap patches are more recent than app
At startup, the instant runtime now looks at the age of the hotswap
patches, if any, and if they are newer than the most recent coldswap
patch (or the APK itself), we display a toast in the running app:

    Your app does not have the latest code changes because
    was restarted manually. Please run from IDE instead

This will hopefully make it obvious what's going on if you manually
run your app for some reason after you've made a hotswap (since
coldswap patches are only pushed on verifier failure or when the app
isn't already running.)

These checks are done early during startup (when we're configuring the
app). At this point it's too early to show a toast (because toasts
require an associated activity), so instead we schedule the toast and
check every second (for up to 10 seconds) for an available activity to
hang the toast off of.

Long toast messages will now also use the long-duration toast instead
of unconditionally using short-duration toasts as before.

Change-Id: I307b58a0d224142850df87940d4c89577ffe35d3
(cherry picked from commit 2f4a0007c5e7a8af443aa7db951697112d6921b8)