Remove the restart-dex support from the runtime library.
This is part of the fix for
204341: Instant Run ClassNotFoundException when switching productFlavors

This CL removes all the support for Dalvik-era "restart dex" patches:
these are coldswap patches that are not sharded, and are instead
sorted by age and then placed in reverse order into the class loader
(such that the newest version of a class is found first).

This code (and in particular the part of that code which was intended
to ignore these patches when the APK had been updated) would wipe out
all the dex slices from the class loader in some scenarios. One
scenario is when there are older slices on disk when there are
multiple differently named slices (and this could happen with
variants, where the SHAs seems to differ by variant).

Even though we're separately fixing the slice naming issue, we might
as well remove this code since it should never legitimately execute
anymore (and if it does, as in this bug, bad things can happen), and
besides, we might as well make the runtime library have as few methods
as possible.

Change-Id: Ic831411d1e3d80e3547e3866a2f7a11a9d6270c4