Bug 27775872: Instant Run broken on latest N build
In recent builds of N, a lot of the internal fields in Resources and Theme
have been moved into separate implementation classes (ResourcesImpl and
ThemeImpl). This requires updating the runtime library to look at some
extra fields when it fails to find references to assets, glyph caches etc
in the normal places.

Change-Id: If2826f70b12f3c03469a8f68c3c91731dbe2f0df