do not process @{applicationId} in libraries.
when doing placeholder replacement in library, in case of
applicationId not provided in the build.gradle, we used the library
implicit package name to replace those ${applicationId} which is
incorrect, we should have used the application's pakage name instead.

this CL will now delay the {$applicationId} resolution until the
final stage of the manifest merging and use the application's
package name instead of the library's.

Change-Id: Ie577034ce427627ed22605f936f4ac5d2b66602a