Add DateProvider to allow stubbing 'now' in tests.
instead of calling "new Date()", code should have a variable initialized
to DateProvider.SYSTEM and call .now() on this instance, tests can then
replace that instance with a custom notion of time. Most commonly
test can replace the instance with a VirtualTimeDateProvider
which will use a VirtualTimeScheduler's notion of time to make tests
have a consistent notion of time. that is in sync with the
VirtualTestScheduler.

Change-Id: I8b33142f14f726a8e04a317a1520c70c43a0132d