IR: maintain a new *.ir.ap_ file with added assets
This change adds a new task which is responsible for keeping an *.ir.ap_
archive up to date. The new archive is similar to *.ap_ but contains
assets and is zipaligned.

This means that the initial (FULL_APK) instant run build is slightly
slower, since we need to create the *.ir.ap_ file from scratch. When
using new packaging (which is the default), the file is updated
incrementally, so we do minimal IO during hotswap builds.

This also fixes the critical bug where hot/cold swaps caused all assets
to disappear from the app.

Bug: http://b.android.com/215196
Change-Id: Ia98792f11010d41adf50e8a62c32c2716421251e
(cherry picked from commit e1b1773abf5033c51f6103837848eea368e19dd8)