Added MergedResourceWriter tracking.
With the introduction of appt2, a resource compiler was
introduced in MergedResourceWriter. However, no tracking had
been added to compiled files, making incremental compilation
incorrect in some situations, e.g., file removes.

This CL fixes this by introducing tracking of files in the
MergedResourceWriter.  ResourceWriter keeps track of compiled
files (when these are different from the original resource
files) and deletes them when the original files are deleted.

Change-Id: I315a07e44b700d66ef0103d28cf0c68c22f70e26