Use "cmd package" instead of "pm" to install apks
APKs are installed using the package manager, which is invoked via the
pm command. This requires spinning up a new VM. For split APKs,
installing a single split causes the creation of 3 VMs (one for each of
session create, write and commit operations)

In API 24 and above, a better scheme is to use "cmd package" which will
run in the system_server process, and significantly improve the latency.

b/30039694

Change-Id: I1979ba25edf5a40f81cc978f635e6f6066a4ef6d