Remove an IDE hack from the main profiler cmake file
The hack added to fix some IDE code completion error should be completly
unneccessary as the include_directories(${CMAKE_CURRENT_BINARY_DIR})
line will have the exact same effect. Additonally that hack can break
the build in very misterious ways when compiling for android after a
change in a proto file without re-generating them in the the host build
as it will cause the compiler to pick up the wrong include file.

Change-Id: Ic44c29c38ff8c982bca43320a437c1ac9d9713c2