instant-run: Always include certain logs to facilitate debugging
Issues like http://b.android.com/215805 could arise because of a number
of reasons. We need more info to debug these scenarios, the very basic
ones being whether the custom BootstrapApplication was triggered or not,
and whether it was able to add dex files to the path. This CL turns on
those two log statements.

Change-Id: I07f5e538b0fbe3010d516810d2d8099024233441