Jack logging
There are 2 aspects of logging in Jack, logging Jack
internals, and logging the compilation/pre-dexing.

When users enabled debug log level, we log the Jack
internals, to help debugging Jack issues.

When info log level is used, we set the Jack reporter
to be more verbose when compiling the source code,
or processing libraries.

Change-Id: I95e282fcf54fc57d97da75ccc97a7e7d3678b8b8
Test: JackCompilationOutputTest