227687: NullPointerException in WrongIdDetector
In addition to fixing this bug, this CL also generalizes the existing
detector protection. There were some known ECJ bugs that would trigger
from time to time, and it would take down the whole build (crashing
lint, then the Gradle build).  To prevent this, lint added a try/catch
around each file visitor such that not only does the build not stop; a
failure to parse a single Java file doesn't prevent checking all the
other files.

This CL generalizes this handling such that it also applies to the XML
visitors. Here's what it now emits for the error above (with the fix
disabled, without the line wrapping) :

Unexpected failure during lint analysis of layout3.xml (this is a bug
in lint or one of the libraries it depends on)
NullPointerException:WrongIdDetector.afterCheckFile(WrongIdDetector.java:228)
<-ResourceVisitor.visitFile(ResourceVisitor.java:160)
<-LintDriver.checkIndividualResources(LintDriver.java:1963)
<-LintDriver.runFileDetectors(LintDriver.java:1155)

Test: Regression unit test included

Change-Id: I8fe459524e0e239f79703f07bf9ce3a335577f09