Added support for symbol table merging.
Removed the old SymbolLoader and SymbolWriter as they are
no longer used.

The current behavior while handling R.java is exactly the
same, with the difference that an intermediate in-memory
representation of the symbol table is now used and the
import / export is now clearly separted from the
in-memory representation.

Change-Id: I2ce44c20656e93389d92fe1030dbfff0c432a848
Test: Included as JUnit tests.