Implemented run-as server for broken run-as devices.
This is an Android Service injected into the AndroidManifest.xml,
which is started on demand when we discover that device has broken
run-as functionality.  The service runs as the app's main user.
We connect to it with a domain socket and send it shell commands.

Test: Manually test cold swap on Samsung Galaxy S4.
Needs a new test framework for unit tests and integration tests.

Change-Id: If46620ae98ed44604d2f1b7532544738fb0372e9