Add serialVersionUID using asm visitor
Previous solution required us to load the class
in order to calculate the serialVersionUID. This
created linking problems, and we were not able
to successfully load the classes from aars containing
signed classes, because the library's R class would not
be signed, and class from classes.jar would be.
Because of this, the SecurityException would be
thrown.

Now we are using ClassVisitor to accomplish the
functionality, and there is no need to load the classes.

Change-Id: Iddc9d100cad560da1a2aad141039b874af739c8b
Test: existing tests in IncrementalSupportVisitorTest
Bug: http://b.android.com/229240, http://b.android.com/225547