Use file path + timestamp + size to identify an aar.
We previously used hashes to identify aars and to compute the locations
of the exploded aars. While this is correct, it slows down the builds
significanly as we frequently need to compute the locations of the
exploded aars.

With this commit, we now use a combination of path, timestamp, and size
to identify an aar. There are 2 side effects:

- Correctness issue: Aars with the same path, timestamp, and size, but
  with different hashes will be considered the same, which is incorrect.
  However, it may be unlikely that this will happen.

- Performance issue: Aars with the same hashes but with different paths
  or timestamps may be cached separately in the build cache, making the
  cache grow larger than necessary.

Bug: http://b.android.com/229654
Test: Updated PrepareLibraryTaskTest
Change-Id: I708860091b34781007f144af0ff818bcdc9eee41