229626: Don't use webp for launcher icons
The developer console requires PNG files for launcher icons.
We recently switched to using WEBP for launcher icons for default
projects if the minSdkVersion >= 18, but that means that as soon
as you try to publish the app it will be rejected.

This CL reverts the templates back to using PNG in all cases
(reverting d4c3013081926d1dd450fe91cd0bb3e861ac0879), and it also
updates the icon detector to flag launcher icons that are in
the wrong format (e.g. WebP or vector drawables, as requested in
issue 229626.)

Test: Unit tests included

Change-Id: Ic14a2e179270ffceab5e663e39a596dcd7b589b0
(cherry picked from commit c1e8e42f7d75578c2c488c312d8ced743bb906b1)