Make plugin check for exact level2 sync.
Previously the code would look for level 2 or
above to do something different.

Since level 2 is not ready yet, but disabling it is non
trivial (since we'd have to rewrite a bunch of tests),
the new code checks for exactly level 2.

We will ensure that Studio never uses level2 and skips directly
to level3.

Test: existing integ tests.

Change-Id: I32cecdca73f5b4fe3d0601c1556451d69a0e0251