throw runtime exception when gradle version is too old.
we used to only generate a sync event when the plugin runs
on a version of gradle which is too old. Unfortuantely, continuing
the sync with an older gradle version can be problematic if the
plugin uses APIs not yet available in that gradle version.

Instead, we geneate an immediate exception that the IDE will
watch and propose to upgradle gradle.

Bug: b.android.com/231658
Test: manual testing.
Change-Id: I2b13887eb291d2732d1ffb814de3198dd17d63bc
(cherry picked from commit 4316ef9d8dbe09658ab7fb4ac47504a3b94103db)