split apk installer: determine success from the first line of response
The split apk installer was hooked up to use the regular package installer's
output receiver. This receiver parsed every line of the output, and if it
did not see "Success" in any of those lines, it marked the installation to
be a failure. This worked because the output was typically just a single
line.

In the case of split apk installs, some devices could return output of the
form "Success: streamed 4754 bytes [CDS]close[0]", and the existing receiver
fails when it sees the second line.

This CL includes a new receiver that follows what adb does: it determines
success based solely on whether the first line starts with "Success".

Test: unit test for new receiver included
Bug: http://b.android.com/232060
Change-Id: I5a274e611bed5bf6ecffb784058772840a4ded5d