Fixes comparision for nextSibling.
Comparison of activityAlias against nextSibling was incorrectly not using the
result of getXml(), resulting in comparison of different types, causing the
crash 'org.w3c.dom.DOMException: NOT_FOUND_ERR' if nextSibling is the same as
activityAlias. Also fixes reordering of <activity-alias> comments to after the
activity alias if they immediately followed the target activity.

Test: new unit tests
Change-Id: I4d3f3c349731dcc8df30be28134881bc8fb248b5