Fix bug 35841603: updating zip with invalid extra
ZFile did not support updating a zip that required aligning
an entry using the extra field if the previous entry had an
extra field that was non-empty but did not match the
format in the spec. With this update, the contents of the
extra field are discarded if they cannot be parsed.

http://b/35841603
Test: Included JUnit.

Change-Id: I977269b0648cd1b3fb7629929e02b9a00f2f0bbf