Add multiple inputs/outputs and in/out dirs for Desugar
Desugar now supports specifying multiple inputs/outputs
in a single invocation. This allows us to save time by
not starting a new JVM for every input/output pair. Instead,
we run Desugar with multiple inputs/output that are sharing the
same classpath, bootclasspath, and other options.

Also, we are not zipping inputs before processing them with
Desugar, as it supports directory inputs and outputs.

Test: existing tests
Bug: http://b.android.com/234217
Change-Id: Iab61194495bf1f1e235c2da32775e5b4c2a97d34