New task output handling.
VariantPublishingSpec provides a declarative model with
bi-directional mapping between a task output and a published
artifact.

Tasks can simply register their output and it'll get published
if it's required.
In the opposite direction, when a test queries for its dependencies,
the tested artifact is automatically added based on the mapping
and the queried Configuration (compile vs runtime).

Test: existing integrations plus some new unit test ones.
Change-Id: I079afc822334fab99c05c3bc3b402f0144490e76