Tune performance of lint annotation check
A snapshot from a customer shows that lint has some hotspots
around directory lookup (which in turn is done to look up
packages and package annotations, as part of the annotation
check).

This CL works around this by providing lighter-weight alternatives.

It also removes the build/intermediates/classes folder from the
source path such that lint doesn't have to consult both the
.java and the .class matches for a given fully qualified name.

Test: Covered by unit tests already

Change-Id: If71bd9acd20aa407d1a4c05bf9cdfddfbdd00f3c