make sure InstantRunBuildContext uses a final isInInstantRun boolean
We used to have a mutable boolean for whether or not we are in
instant run mode in the context. It used to work fine until aapt2
integrated (I believe) and the mode was not initialized soon enough
leading to not packaging the right resources with the content provider.

made the mode a final variable of the context and adapted existing code.
modified an existing integration test to make sure IR ContentProvider
has been added to the manifest file.

Bug: 37476163
Test: added integration test verification.
Change-Id: If6c20d2518ad909aee460bf67dc5446bdb1b4614