Add an application dependency scope for features.
This ensures the installed application and the instantApp share the
same packageId.
This is achieved by modifying the configuration used to publish the
feature manifest elements to a more generic "metadata" configuration.
This configuration is used by the application plugin to publish its
packageId, which is eventually consumed by the base feature through the
application dependency scope.

BUG=36530655
Test: Existing

Change-Id: I944169d17faeb9a8a30e7cb64d0dabc4fcf4e4d0