Fix an issue with preBuild for androidTest variant.
Even though the requested ArtifactCollection query for external
only dependencies, the task actually used sub-projects dependencies
as well due to how FileCollection.minus works.

This led to crashes after we switched from using CLASS_DIRECTORY
to a custom CLASSES artifactType because this forces sub-projects
to go through an ArtifactTransform (to go from their jar output
to CLASSES). However because the sub-projects were not built
yet this failed.

The issue around minus is due to the fact that it works strictly
on files, meaning that it has to resolve both file collections
fully and then substract the list of files.

The collection built for the test runtime classpath went through
VariantScopeImpl.handleTestedComponent which removes the tested
runtime classpath in order to de-duplicate dependencies. However,
this was done on the ALL scope always instead of matching the
queried scope, adding files we didn't care about (and couldn't
handle).

Even trying to properly setup the task dependency to build
the sub-project doesn't work, as the task dependency mechanism
using @InputFiles requires querying the file collection in the
first place, which cannot happen until the artifacts are built
and the ArtifactTransform can run. Kind of a chicken and egg
problem.

This changeset also properly sets up TestPreBuildTask to
depend on the artifacts it needs and to have a fake output
folder to be UP-TO-DATE unless the dependencies change.

Change-Id: Id11a39b593792e70ba24a13f67043ef62c5681b0
Test: need to add new ones to cover this but at least this broke the perf projects so it's covered, sort of.