Refactor CPU profiler type to avoid duplicated fields
As we add support for simpleperf, CPU profiler type becomes an important
field, because we need to start adding checks for it. For instance, the
trace parser used depends on the profiler type.

For that reason, we should avoid repeating the same enum inside multiple
proto messages. Instead, we should provide a single entry point.

Test: existing tests
Bug: 62215417
Change-Id: I9f8ca92d3ddca2c2b823d70b22219965d4d427d1