Remove Kotlin unused symbol/import inspections.
...until entry points are defined properly. Before enforcing Kotlin
inspections on PSQ, these ones need to be removed from the profile as
they may report false positives since a lot of stuff in tools/base is
used outside of the particular IDEA project. Long term fix is to 
configure a set of entry points so IDEA doesn't get confused.

Test: N/A
Change-Id: Id7462ca28690334fb4390d9dd696cfe338cde79f