Merge and renumber the IDs generated for library projects.
This fixes some issues when not running the resource merger
in library projects as we have to merge the R.txt manually.
(this is behind a flag for now)

IDs in the R.txt files are generated in the PPTTNNNN format
of aapt, however it seems like different versions of the plugins
have used different values (TT) for each ResourceType enums.

This leads to conflict where different resources of different types
share the same ID, and that breaks tools that parses these, like
robolectric.

The change does a merge + renumber as a single pass, in order
to properly handle styleable arrays that reference ATTR values.
The previous merge would potentially erase values and prevented
resolving array values back to attr which prevented renumbering
them.

Also fixes an issue in the resource parser that generates the
IDs for the current module where the value in the array was confused
with the index values.

Finally, this also removes the ability to disable package
name enforcement (where a library cannot be the same name
as the main package) as all this work isn't really compatible
with it anyway.

Change-Id: I7a34c5f09578cf89373156688034ff67f779a194
Test: existing tests.