[AVD Mgr] Set screen size for skinless AVDs
Most AVDs have a skin that indicates the screen size. When creating
an AVD from the command line, the user can indicate a reference device
without using that device's skin. If we don't retain the screen size
from the reference device, the Emulator will use a very small default
size.

Note that the Emulator prefers the value specified by the skin, so this
change has no effect when a skin is used.

Bug: 62905584

Test: Updated DeviceManagerTest tests and added .testGetHardwareProperties

See also https://android-review.googlesource.com/422264

Change-Id: I8b008a690120b08e2c075eb0a8205d51fdd7ddbc