Fix studio model for local jars inside AARs.
The integration test verifying this was update to a newer
version of the support library that does not contain any
local jars inside its AARs but the test was not designed to
detect this.

Fix the test, using an older version and made it expect a certain
local jar.

Then fixed the code to remove the FIXME! with actual compuation
of the local jar list.

Also fixed the test to test both L3 (currently used by Studio
3.0-a8), and the upcoming L4.

Bug: 63948066
Change-Id: Idfec8217b53bdd2d0d2112be2d39b6e2df580cf6
Test: LocalJarInAarInModelTest