Support lambda captures in call graph analysis
Specifically, when exact types flow into variables captured by
a lambda, that lambda becomes specialized on those types and can
itself flow through the program.

Invocations of captured functional variables in Kotlin is not supported
yet. It's better to wait until the Kotlin UAST issues can be fixed, as
the required workarounds would pollute the code too much in this case.

Test: CallGraphTest
Change-Id: Ic64b6dbd7f14bfa9f03d84f587374b70f5c185ca