Fix kotlin-related error with ide model builder.
There are two kotlin build output test class directories
specified, but the ModelBuilder is only set up to handle one.

The way build output dirs are handled is changing in beta3,
and this code will no longer be called, so this is a temporary
workaround for the P0 project creation bug.

Fixes: 64428659
Test:  Fixed ModelBuilderTest

Change-Id: I4c25581f552ec854b90f9ee8147ce30e89a9415f