Add a <buildtype>AndroidTest sourceset.
This also adds the missing androidTest<BuildType>Implementation
configuration object.
Without this, any dependencies added to debugImplementation is not
inherited by the androidTest variant. This is a problem since we are
now relying on this configuration extension mechanism to set the
dependencies of the tests.

(In 2.3 we did this manually after resolving the tested dependencies
during config).

Bug: 65505901
Change-Id: Ib79120f156d935e1dd9297e2f35963fe2842fba9
Test: Updated ConfigurationExtensionTest plus other tests