Split up ProcessAndroidResources task, step 1.
As it has too many responsibilities.

It is moved to LinkApplicationAndroidResourcesTask, and a new
GenerateLibraryRFileTask is extracted from it.

In order to preserve API compatibility,
ProcessAndroidResources is renamed LinkApplicationAndroidResourcesTask,
and both LinkApplicationAndroidResourcesTask and
GenerateLibraryRFileTask extend from a new abstract
ProcessAndroidResources.

Checking that the injected ABI is present is moved to the
ExternalNativeJsonGenerator, so it is now checked at configuration time.

Test: existing gradle integration tests.
Change-Id: I3a0eec005b5de9edd415307b8cad9943fac14b3e