Unified Android model.
This is a reworking of the model in builder-model. The intent is to:
- Make the model easier to instantiate (by replacing interfaces with
  Kotlin data classes that supply defaults for most attributes).
- Make the model easier to use. The public contracts are designed from
  the perspective of simplifying the code that uses them rather than
  optimizing over-the-wire performance.
- Remove API that isn't used by the IDE or linter.
- Generalize the model so that it applies equally well to gradle and blaze.
- Remove deprecated API that is only used by old versions of Gradle.

This is intended to be used with builder-model - the
Gradle-specific code will continue to use builder-model, which is
optimized for interprocess data exchange and handles backwards
compatibility with older gradle versions. After sync, this model will
be constructed and handed off to the rest of the IDE for consumption.

Design doc: go/android-project-model
See project-model-readme.md for more information.

Bug: 68848940
Test: Created android-project-model/testSrc with several test suites
Change-Id: Ie188c97fff582186b20d38cba022f35bf092e951