Fix issue with workaround for Gradle's @Optional bug.
If an @Optional @OutputFile/@OutputDirectory property is set to null in
a build, subsequent builds will always be  UP-TO-DATE even if the output
file/directory is set to a different value.

In commit e022bb88257b2a6c42256b1ade8e25aabbb5256f, we worked around
this Gradle @Optional bug by combining the paths of those output
files/directories and mark it as an @Input.

However, this also means that if an output path is changed from a
non-null value to another non-null value, the inputs are also changed,
and the Gradle cache cannot be used even if the other inputs remain the
same.

To work around the @Optional bug but still ensure the relocatability of
caching, it is sufficient to combine the nullability of the output paths
instead of the paths, so that the inputs are changed when an output path
is changed between null and non-null, but not when it's changed between
non-null values.

Test: Existing MergeResourcesTest, Updated CombinedInputTest
Bug: 68771542, 67418335, 65829618
Change-Id: I372b302c790420d48e92ce596b3383872a3cc443