Change version of dependency in Uber project
Gradle 4.4 resolves optional dependencies, what is causing conflicts
between com.android.support.support-annotations[24.0,25.0) (which is an
optional dependency of com.braintreepayments.api:braintree:2.3.6) and
other versions of the dependency.

While this bug is not completely fixed, this CL changes the version of
braintree dependency so there is no conflict.

This CL should be reverted if optional dependencies resolution is fixed
(filed bug 68993591).

Test: UberSkeletonSmokeTest
Change-Id: I4f6560844f4375b25b445f2b60234be85930ab3e