Refactor the AntennaPod performance tests.
The purpose here is to create the concept of a "benchmark" that can be
runs in a hermetic fashion. The plan being to eventually shard this
work over a number of slave machines.

cmw@ suggested that this would be a good Kotlin learning experience, and
that parts of the code lend themselves nicely to being written in
Kotlin. As such, most of this code is Kotlin.

I haven't written any Kotlin before this, so please review thoroughly.
:)

Test: AntennaPodPerformanceMatrixTest
Change-Id: I24f8e0125237179816e4408f584a04d99c47f415