Create classpaths configurations in a constant order
This replaces the collection holding the classpaths from a HashSet to
a LinkedHashSet. HashSet makes no guarantee on the order of elements,
unlike LinkedHashSet. In addition, next is O(1) on LinkedHashSet
instead of O(h/n) on HashSet. This comes at a small increase in memory
footprint (4 bytes per classpath), but the set of classpaths is small.

This fixes a bug where incoming artifacts would come in a different
order across builds, making incremental builds not up-to-date when
they should have been.

BUG=67041434
Test: Not consistently reproducible so N/A

Change-Id: Ia508b56124db930c232850f8955a44af6197b651