Added StringOption to set location of benchmark outputs
Profiler .rawproto outputs will now no longer be generated unless
PROFILE_OUTPUT_DIR is set (using either a path relative to the project
root or an absolute path). If ENABLE_PROFILE_JSON is set but the new
StringOption is not, the outputs will be generated in
build/android-profile as a default.

Benchmark tests must use GradleTestProjectBuilder.enableProfileOutput to
set this flag so the ProfileCapturer will be able to get find the
outputs. This method also defaults to build/android-profile, or the
output can be changed with enableProfileOutputInDirectory()

This also solves the issue of files existing in the build/ folder after
a clean, because the files are no longer created by default.

Bug: b/37129229
Test: Existing tests
Change-Id: Ide7dda2d89df4f73183a2f9eefb4a7d1cfac492c