Configure kotlin kapt for data binding
This CL changes task manager to detect the existence of kotlin
plugin and configure data binding dependencies for both the
kapt configuration and kapt tasks (kotlin compilation).

This is a bit hacky, normally such thing should happen in an
API between AGP and kotlin plugin but this quickly fixes the
current issue and we can iterate over it.

Bug: 70915745
Test: DataBindingKotlinAppTest
Change-Id: I73f0b619bffed71feced91dc637313a1d48b13d5