Rules parser: accept types before <fields> and <methods>
It's not clear what such rules mean exactly, I experimented with it a
bit but it doesn't cause all fields/methods with the given return
type to be kept.

Still, ProGuard accepts them and it seems we must as well, since
sometimes people use this syntax in their flags. For now we'll just
ignore the type and keep all fields/methods, which is the safer way
around.

Bug: 70841987
Bug: 67835701
Bug: 71796333
Test: ProguardParserTest
Change-Id: I4a985510deafa75e78c70de968f62f6e17acbb8a