Pull profile uploading out of benchmarks.
Historically each Benchmark has been responsible for uploading its
benchmark result. This feels like an inversion of control, and the
top-level class, BenchmarkTest, should be in charge of it.

Moving this functionality makes the code easier to reason about and
re-use, as well as paving the way for making it easier to test
when we change BenchmarkTest to be a simple wrapper around some
BenchmarkRunner class that has all of its dependencies injected.

Test: bazel test -- //tools/base/build-system/... -//tools/base/build-system/integration-test/...
Change-Id: I91e05e985f8acc20f0f62b3316ddd175d3d7b4e1