New stub gradle plugins in gradle-api.
This moves the gradle-api artifact closer to being the only thing
required for gradle plugins that interact with the android plugin to
build against.

They are all marked as incubating. The existing plugin names, e.g. as
used in "apply plugin: 'com.android.application'" now point to the API
plugins (in com.android.build.api.plugin). Those plugins then apply the
internal plugins.

The internal plugins also apply the external ones to avoid confusion if
the internal plugin is being applied by class name. This is safe as
plugin application is idempotent.

Eventually, we should consider moving the internal plugins to make it
clear that they are not API.

Test: existing, StableApiTest.
Bug: 79705349
Change-Id: I97a15ea8f0121f665bb740d3d5e68e6608ccce51