Sets property required for running on Foundry
It seems that when we run with a different JDK we need this
property set. This property is read in a static initializer
of a class, and it's unclear when this class is loaded.
This issue used to happen when running locally too, and it
went away magically. Well, it didn't.

Test: this is it
Bug: 78922354
Change-Id: I222f60211841b784f6ec3e48aa498d7e17c77a90