Change the format of RDef.
Use the symbol table with package name format. Decoupling it from the
partial R files, as it will contain xml (non-canonical) resource names
in a subsequent change.

Write a header that the reader asserts is present to ensure that the
correct parsing function is used.

Fix up SymbolTablePartialRMergeTest that was using it as a fancy
toString() method, and LibraryDefinedSymbolTableTransformTest which
asserted directly about the content for the file.

Bug: 73626400
Test: SymbolIoTest.
Change-Id: Icf9cc209ca3f18dfc34de3a8d2a501f6255d9359