Fix CMake server mode in the presence of Hunter module
This may or may not be Hunter specific. The issue is that CMake
server mode can return a FileGroup that has includes (and defines
by the way) pulled into a separate field. It's not clear when
this happens and when it doesn't. This change pulls those
includes and defines back into the compile flags so that they
can be properly transmitted up to Android Studio.

Bug: 80513018
Test: Added a section in NativeModelTest
Change-Id: I201b18ccaf392046441a5dc9ae6c652ef5c39db9