Add toString methods to more data objects
Config and ManifestAttributes previously had an inline algorithm for
printing data objects via reflection in a way that omitted optional
attributes that were left with their default values. This code has been
generalized, moved into DataObjectUtil, and used to provide better
toString methods for AndroidProject, Artifact, ArtifactDependency,
and Variant.

This change is mainly intended to help with viewing models in the debugger.

Test: ConfigTest, ManifestAttributesTest
Change-Id: I637f4255a5075ef5bf167e1a0025922efa9d6b2d
Signed-off-by: Stefan Xenos <sxenos@google.com>