Add mechanism to run @Ignore'd tests
This change provides a mechanism to run ignored unit tests, so their
status can be monitored without enabling them on and thus disrupting
the main builds.

This is achieved by adding a bazel test configuration called 'ignored',
which sets the ignored.tests.only system property. JarTestSuiteRunner
then substitutes IgnoredTestsRunnerBuilder (defined in this CL) for
the default one when that property is set. This RunnerBuilder attempts
to create Runners that invert the logic of @Ignore as closely as
possible. There are a couple of limitations:

  * Tests that use @RunWith to run with custom Runners can't be handled
    in general, since it's not possible to know what the custom Runner
    does. A couple of important cases, Suites and @RunWith(JUnit4.class),
    are handled specially. Other special cases could be added if
    necessary. The builder creates IgnoredClassRunners, which don't run
    anything, if it doesn't know what to do. The most common unhandled
    case is parameterized tests.

  * For JUnit3 tests, @Ignore annotations are limited to class-level
    granularity. @Ignore has no effect when applied to JUnit3 methods.

To run ignored tests, run

    bazel test --config=ignored <target>

If need be, tests can be completely disabled by removing the @Test
annotation (for JUnit4), or removing the 'test' prefix on the method
name (Junit3).

Change-Id: I6d9e18c742b2559cb65aab3af77d77ce0d9ae6fc