Preserve targetSandboxVersion if manually set in manifest.
targetSandboxVersion was getting overwritten for instant app manifest purposes,
then the assumption was that targetSandboxVersion was not present in the
manifest, so it was removed to undo the change. This detects if
targetSandboxVersion was set, and restores its previous value if so.

Bug: 117176166
Test: Added tests to ManifestMerger2SmallTest
Change-Id: If99d141089f8dfd05d8c7da25722b6b22defc74e
(cherry picked from commit 19a81d73cef364be16539d43fb6f657cb317a621)