Move the package name from Config to AndroidSubmodule
The package is often needed for arbitrary SourceSets - even ones
belonging to Configs that don't specifically overload the package -
and it is required to be the same for all Artifacts, Configs, and
SourceSets in the same AndroidSubmodule.

Moving it to AndroidSubmodule makes it available everywhere it is
needed and correctly reflects the real invariants.

Test: AndroidSubmoduleTest
Change-Id: I9e1106c2f18224845e7fce73486a3c849cfd58c6