IRR: Fix bug 117904350
This bug happened because some event on the agent could remain in
the agent process after a swap request. The next swap request then
picked up this event which as used as the earliest remove event,
even thought it was issued in a previous run.

The fix is to purge the event system when the Event System is
initialized in the agent process, on each swap request (the entry
point is Agent_OnAttach).

Bug: 117904350
Test: Manual
Change-Id: I162394162f564cdb773a061e39f40c0906133233