Incremental desugaring - cleanup previous state
This CL fixes cleanup for incremental desugaring
with D8. Previously, for a A.jar, we could end up
with A.dex copied from cache and with sharded A dex
files i.e. A.dex-1, A.dex-2... in the output.

This CL fixes the cleanup, and for every jar that
needs to be desugared+dexed because some of its
dependencies changed, we cleanup the previous state
correctly.

BUG: 117062425
Test: DexArchiveBuilderTransformDesugaringTest
Change-Id: I1ef22febc3e02df8b58e33733e855ef1ec5542a2
(cherry picked from commit d14e056d3c94ac312078d9a42e358082d1b2b0d9)