Fix the kotlin stdlib version inconsistency
ktx library depends on specific kotlin-stdlib library, but it results in
a conflict as in the bug such as in a instant app project:
======================
Conflict with dependency 'org.jetbrains.kotlin:kotlin-stdlib' in project ':app'.
Resolved versions for runtime classpath (1.2.60) and compile classpath (1.2.50) differ.
======================

This CL is to add the kotlin-stdlib dependency to the monolithic app
module.

Bug: 112764077
Test: Manually verified by creating an instant app (needed to change the
      UAB_NEW_PROJECT_INSTANT_APP_IS_DYNAMIC_APP flag to false)
Change-Id: Idcb807af0042c305f90b1ef6a491f8d85355f888