Ensure cleanup after cancelled SDK installation.
When an install operation is performed, it typically means unzipping
into the target location under Android SDK tree, which is not an atomic
operation. If it gets cancelled during unzipping, the tree is left in
an inconsistent state, which may confuse certain package locator helpers
and in general such a situation is better to be avoided.

Bug: b/68788981
Test: Added BasicInstallerTest.testCleanupWhenCancelled
Change-Id: I37291a7753f98d6cd4e17341d6754d4aa1c9d831
(cherry picked from commit 0f0a3b9f536cc38a4674135c8441532244666551)