Remove workaround for incremental compile before Gradle 5.0.
There was a bug before Gradle 5.0 that if a non-incremental annotation
processor changes a resource, Gradle will mistakenly remain in
incremental mode, and may not even perform any recompilation at all even
though the JavaCompile task is not UP-TO-DATE.

We had to explicit disable incremental mode in that case, to force
Gradle to perform a full recompile.

Now that we have updated to Gradle 5.0, let's remove this workaround.

Performance impact: The workaround may have worsen incremental compile
time for non-ABI changes in 3.3 (compile tasks are out-of-date even when
they can be UP-TO-DATE), but the workaround was needed for 3.3 for
correctness (it was too late for 3.2).

Removing this workaround means that incremental compile time for non-ABI
changes will improve, it will be at least as fast as in 3.2.

Bug: 113054294
Test: Existing IncrementalJavaCompileWithAPsTest,
      DataBindingIncrementalTest

Change-Id: Ib1c09724c0b46f30eee2e2a15ab62d4ba6be2f74