Workaround for Gradle 5.0 and kotlin lateinit
With Gradle 5.0, Provider<> @InputXXX fields will be queried before the
task is initialized leading to a lateinit field query before the field
as been actually initialized.

Force the initialization to a useless value to satisfy Gradle 5.0
The value will be overriden as soon as the init block runs.

Reported by Gradle in https://scans.gradle.com/s/62bxxdf6kpbc2

Test: None
Bug: N/A
Change-Id: I74f7437e7b24f2f447b954fbf45467695bd93909