3.3 issue: Windows slash in resolve C++ file flags
User was reporting lost code completion in their project. Issue was
that lookup of flags with key of filename traveled through File(xxx).path
in one case and raw string in another case.

This was a problem only on windows because CMake emits paths in the form:

      C:/path/to/my.cpp

The File(xxx).path of this is C:\path\to\my.cpp only on Windows. But the
map of file to flags was keyed by C:/path/to/my.cpp so the lookup
failed.

An additional problem with the code that is note fixed in this CL is that
when the lookup fails the code tries a second method to find flags which
is known to be wrong (at least sometimes). I didn't want to turn this
into an exception so late in 3.3 cycle so I left it as-is for now.

Bug: 112611156
Test: Added unittest for it. Test would only have failed on Windows.

Change-Id: Ib3b0b56111fc9ecbe409700d87a0f4127d0d1d3e
(cherry picked from commit 093e08e0f2cd0670b6bf26c76f3312c67e691579)