Compiler settings cache: function to interpret CMake boolean property
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

Adds a function that interprets a CMake property definition for
true or false boolean value. Function name uses the term "truthiness"
from SQL domain to describe a string value that is interpreted as
a boolean.

Bug: 119116588
Test: NA
Change-Id: I3dcd99c077894f411ded8f6ef6fbdcf4944fcf06