119220179: New lint check for accidental SAM conversions
This CL looks for cases where you're passing a lambda
expression (or callable reference expression) to a
Java method which takes a SAM parameter, and then that
method does something with the parameter such as storing
it in a list or doing an instance compare; this can lead
to subtle bugs. We had a memory leak in Studio as a result
of this, where we were passing in a listener to a method
which registered and unregistered Runnables. Each call
created a new Runnable, which meant remove listener did
not infact remove the earlier registered instance.

Test: Unit test included
Fixes: 119220179
Change-Id: I9eaeb28eb1dc3394b6634b509087b0905bf9c837